Add New Styles for Updated ODT files.
New styles have been added and changed in the template files. Add
new styles to accommodate these updates. Also add a test which
verifies the updated styles are found and properly marked as mother
tongue.

Also the invisible style is in a different nesting the DOM so a
special case needed to be added for this style.