Non-Translatable Lessons for full-quarter files
Lessons can now be uploaded which are not to be
translated, but which can receive translated string
substitution. This is useful for full Quarter files.
These can be uploaded and once each lesson has been
translated, they can be downloaded and receive all
the translations -- but don't require any work to
assemble them later on.

This required changes to the API ("admin lessons")
to minimize changes with the desktop application.

Also allow for Covers to be uploaded (allow for the
metadata required, and the UI options to declare
uploads covers along with TOCs).

Plus other minor changes and units updates.